Create Index
I kept the index page content to a minimum and included mainly images because I people get turned off by a too mush content and images leading to people logging off not having looked at the other pages. In relation to my website, It is very attractive and well laid out with picturesque and high quality images and interesting and relevant content underneath the images with the option of reading more about a topic by clinking onto links. I tried keeping the index page as small as possible as I did not too much scrolling to be taking place as again this would turn people off.