Create Places to stay
With the “Places to Stay” page, I had a number of hotels and B&B’s which I structured in separate columns. As already mentioned, I had a number of each hotel and B&B so I did not want to user having to scroll the main page to search for what they were looking for. I made separate scrolls for Hotels and B&B so that when you want to scroll down one of them it does not take away from the main page