Create Blarney
This is a guide into Blarney. When you hover over the smaller images it shows up in larger image. I provided detailed content about the places in Cork to visit. In addition, along the left hand side I gave the option of other places to visit and links to these pages. I feel that this was important as users might not know these places exist and the images and content will draw the users to have a look at these other sites that they may be enticed to visit. This also makes navigating around the site easier because they do not have to go up to the navigation bar every time or down to site map.